Dashboard page: unit tests separated in different files
